Improve file backup.

+++ install.sh
@@ -4,37 +4,33 @@
 # Post-installation script for Debian stable
 #
 
-# Variables
-CONFFILES='/etc/inputrc /etc/apt/sources.list /etc/bash.bashrc /etc/nanorc /etc/skel/.bashrc /etc/skel/.profile'
-INSTALLDIR='/usr/share/debian-postinstall'
 
 # Create installation directory
+INSTALLDIR='/usr/share/debian-postinstall'
 mkdir $INSTALLDIR
 cp -Rfv . $INSTALLDIR
 
 # Backup previous configuration files
+CONFFILES=`find config -type f | sed 's:^config::'`
 mkdir $INSTALLDIR/backup
 for FILE in $CONFFILES; do
-	cp -fv --parents $FILE $INSTALLDIR/backup
+	if [ -f $FILE ]; then
+		cp -fv --parents $FILE $INSTALLDIR/backup
+	fi
 done
 
-# Copy configuration files
+# Backup and remove previous bash configuration files
+DIRLIST="/root /etc/skel `ls /home | sed 's:^:/home/:'`"
+for DIR in $DIRLIST; do
+	cp -fv --parents $DIR/.bashrc $DIR/.profile $INSTALLDIR/backup
+	rm -fv $DIR/.bashrc $DIR/.profile
+done
+
+# Copy local configuration files
 cp -Rfv config/* /
 
-# Remove previous bash configuration files
-rm -fv /etc/skel/.profile
-rm -fv /etc/skel/.bashrc
-rm -fv /root/.bashrc
-rm -fv /root/.profile
-for USER in `ls /home`
-do
-	rm -fv /home/$USER/.bashrc
-	rm -fv /home/$USER/.profile
-done
-
-# Disable network interfaces in udev
+# Remove previous network interface detection in udev
 rm -fv /etc/udev/rules.d/70-persistent-net.rules
-: > /etc/udev/rules.d/75-persistent-net-generator.rules
 
 # Update packages
 apt-get update
